+++ JwtAuthenticationApi/Factories/Wrappers/MutexWrapperFactory.cs
@@ -5,7 +5,7 @@ namespace JwtAuthenticationApi.Factories.Wrappers
     /// <summary>
     /// Implementation of <see cref="IMutexWrapperFactory"/> interface. Factory class for <see cref="IMutexWrapper"/>.
     /// </summary>
-	public class MutexWrapperFactory : IMutexWrapperFactory
+	public sealed class MutexWrapperFactory : IMutexWrapperFactory
     {
 	    /// <inheritdoc/>
 		public IMutexWrapper Create(bool initiallyOwned, string name)