+++ JwtAuthenticationApi/Security/Password/Salt/SaltService.cs
@@ -14,10 +14,13 @@
 	/// <inheritdoc/>
 	public sealed class SaltService: ISaltService
 	{
+		private const int SemaphoreInitialCount = 1;
+		private const int SemaphoreMaximalCount = 1;
+
 		private readonly IPasswordSaltContext _context;
 		private readonly IGuidWrapper _guidWrapper;
-		private readonly IMutexWrapperFactory _mutexWrapperFactory;
 		private readonly IPollySleepingIntervalsFactory _pollySleepingIntervalsFactory;
+		private readonly ISemaphoreWrapperFactory _semaphoreWrapperFactory;
 		private readonly ILogger _logger;
 
 		/// <summary>
@@ -25,26 +28,29 @@
 		/// </summary>
 		/// <param name="context">Password salt database context.</param>
 		/// <param name="guidWrapper">Guid wrapper.</param>
-		/// <param name="mutexWrapperFactory">Mutex wrapper factory.</param>
 		/// <param name="pollySleepingIntervalsFactory">Factory for sleeping interval used in Polly.</param>
+		/// <param name="semaphoreWrapperFactory">Semaphore wrapper factory.</param>
 		/// <param name="logger">Logger.</param>
 		public SaltService(IPasswordSaltContext context, IGuidWrapper guidWrapper,
-			IMutexWrapperFactory mutexWrapperFactory,
-			IPollySleepingIntervalsFactory pollySleepingIntervalsFactory,ILogger logger)
+			IPollySleepingIntervalsFactory pollySleepingIntervalsFactory, 
+			ISemaphoreWrapperFactory semaphoreWrapperFactory,
+			ILogger logger)
 		{
 			_context = context;
 			_guidWrapper = guidWrapper;
-			_mutexWrapperFactory = mutexWrapperFactory;
 			_pollySleepingIntervalsFactory = pollySleepingIntervalsFactory;
+			_semaphoreWrapperFactory = semaphoreWrapperFactory;
 			_logger = logger;
 		}
 
 		/// <inheritdoc/>
 		public async Task<string> CreateAndSaveSaltAsync(UserModel user, CancellationToken cancellationToken = new CancellationToken())
 		{
-			IMutexWrapper mutexLock = _mutexWrapperFactory.Create(true, user.Id.ToString());
+			// IMutexWrapper mutexLock = _mutexWrapperFactory.Create(false, user.Id.ToString());
+			ISemaphoreWrapper semaphore = _semaphoreWrapperFactory.Create(SemaphoreInitialCount, SemaphoreMaximalCount, user.Id.ToString());
 			_logger.Warning($"Trying to acquire lock in {nameof(CreateAndSaveSaltAsync)} method with lock id: {user.Id}.");
-			mutexLock.WaitOne();
+			semaphore.WaitOne();
+			// mutexLock.WaitOne();
 			_logger.Warning($"Lock acquired in {nameof(CreateAndSaveSaltAsync)} method with lock id: {user.Id}.");
 			var dbSavePolicy = Policy
 				.Handle<DbUpdateException>()
@@ -65,7 +71,7 @@
 			}
 			finally
 			{
-				mutexLock.ReleaseMutex();
+				semaphore.Release();
 				_logger.Warning($"Lock released (Lock id: {user.Id}) in {nameof(CreateAndSaveSaltAsync)} method.");
 			}
 			return salt;
@@ -74,28 +80,38 @@
 		/// <inheritdoc/>
 		public async Task<Result<string>> GetSaltAsync(UserModel user, CancellationToken cancellationToken = new CancellationToken())
 		{
-			IMutexWrapper mutexLock = _mutexWrapperFactory.Create(true, user.Id.ToString());
+			ISemaphoreWrapper semaphore = _semaphoreWrapperFactory.Create(SemaphoreInitialCount, SemaphoreMaximalCount, user.Id.ToString());
 			_logger.Warning($"Trying to acquire lock in {nameof(GetSaltAsync)} method with lock id: {user.Id}.");
-			mutexLock.WaitOne();
+			semaphore.WaitOne();
 			PasswordSaltModel passwordSaltModel = null;
 			try
 			{
 				_logger.Warning($"Lock acquired in {nameof(GetSaltAsync)} method with lock id: {user.Id}.");
-				
-				var policy = Policy.Handle<Exception>()
-				.WaitAndRetryAsync(_pollySleepingIntervalsFactory.CreateLinearInterval(2, 2, 3),
-					(x, z) => _logger.Error($"Error occurred during execution of {nameof(GetSaltAsync)}. Attempting to retry in {z.Seconds} seconds. Error Message: {x.Message}."));
 
-				var a = await policy.ExecuteAndCaptureAsync(
-					async () => await _context
+				var policy = Policy.Handle<Exception>()
+					.WaitAndRetryAsync(_pollySleepingIntervalsFactory.CreateLinearInterval(2, 2, 3),
+						(x, z) =>
+						{
+							_logger.Error(
+								$"Error occurred during execution of {nameof(GetSaltAsync)}. Attempting to retry in {z.Seconds} seconds. Error Message: {x.Message}.");
+							_logger.Error($"#ManagedThread# - {Thread.CurrentThread.ManagedThreadId}");
+						});
+
+				var a = await policy.ExecuteAsync(
+					async (ct) => await _context
 						.PasswordSalt
-						.FirstOrDefaultAsync(u => u.UserId.Equals(user.Id), cancellationToken).ConfigureAwait(false))
-					.ConfigureAwait(false);
-				passwordSaltModel = a.Result;
+						.FirstOrDefaultAsync(u => u.UserId.Equals(user.Id), ct)
+						.ConfigureAwait(false), cancellationToken, false).ConfigureAwait(false);
+				passwordSaltModel = a;
+			}
+			catch (Exception ex)
+			{
+				_logger.Error(ex,$"Error occured during receiving password salt for user {user.Id}.");
+				passwordSaltModel = null;
 			}
 			finally
 			{
-				mutexLock.ReleaseMutex();
+				semaphore.Release();
 				_logger.Warning($"Lock released (Lock id: {user.Id}) in {nameof(GetSaltAsync)} method.");
 			}
 			if (passwordSaltModel is null)